script update

+++ script-extensions/configure-win-10-standard.ps1
@@ -1,4 +1,4 @@
-param($sourceFileUrl="", $destinationFolder="", $labName="")
+param($sourceFileUrl="", $destinationFolder="", $labName="", $installOptions="")
 $ErrorActionPreference = 'SilentlyContinue'
 
 if([string]::IsNullOrEmpty($sourceFileUrl) -eq $false -and [string]::IsNullOrEmpty($destinationFolder) -eq $false)
@@ -17,31 +17,43 @@ if([string]::IsNullOrEmpty($sourceFileUrl) -eq $false -and [string]::IsNullOrEmp
 }
 
 
-# Install Chrome
-$Path = $env:TEMP; 
-$Installer = "chrome_installer.exe"
-Invoke-WebRequest "http://dl.google.com/chrome/install/375.126/chrome_installer.exe" -OutFile $Path\$Installer
-Start-Process -FilePath $Path\$Installer -Args "/silent /install" -Verb RunAs -Wait
-Remove-Item $Path\$Installer
+if([string]::IsNullOrEmpty($installOptions) -eq $false) 
+{
 
 
-# Install VS Code
-$Path = $env:TEMP; 
-$Installer = "vscode.exe"
-Invoke-WebRequest "https://go.microsoft.com/fwlink/?Linkid=852157" -OutFile $Path\$Installer
-Start-Process -FilePath $Path\$Installer -Args "/verysilent /MERGETASKS=!runcode" -Verb RunAs -Wait
-Remove-Item $Path\$Installer
+    if($installOptions.Contains("Chrome")) 
+    {
+        # Install Chrome
+        $Path = $env:TEMP; 
+        $Installer = "chrome_installer.exe"
+        Invoke-WebRequest "http://dl.google.com/chrome/install/375.126/chrome_installer.exe" -OutFile $Path\$Installer
+        Start-Process -FilePath $Path\$Installer -Args "/silent /install" -Verb RunAs -Wait
+        Remove-Item $Path\$Installer
+    }
 
+    if($installOptions.Contains("VSCode")) 
+    {
+        # Install VS Code
+        $Path = $env:TEMP; 
+        $Installer = "vscode.exe"
+        Invoke-WebRequest "https://go.microsoft.com/fwlink/?Linkid=852157" -OutFile $Path\$Installer
+        Start-Process -FilePath $Path\$Installer -Args "/verysilent /MERGETASKS=!runcode" -Verb RunAs -Wait
+        Remove-Item $Path\$Installer
+    }
 
-# Install Azure CLI 2
-$Path = $env:TEMP; 
-$Installer = "cli_installer.msi"
-Write-Host "Downloading Azure CLI 2..." -ForegroundColor Green
-Invoke-WebRequest "https://aka.ms/InstallAzureCliWindows" -OutFile $Path\$Installer
-Write-Host "Installing Azure CLI from $Path\$Installer..." -ForegroundColor Green
-Start-Process -FilePath msiexec -Args "/i $Path\$Installer /quiet /qn /norestart" -Verb RunAs -Wait
-Remove-Item $Path\$Installer
+    if($installOptions.Contains("CLI")) 
+    {
+        # Install Azure CLI 2
+        $Path = $env:TEMP; 
+        $Installer = "cli_installer.msi"
+        Write-Host "Downloading Azure CLI 2..." -ForegroundColor Green
+        Invoke-WebRequest "https://aka.ms/InstallAzureCliWindows" -OutFile $Path\$Installer
+        Write-Host "Installing Azure CLI from $Path\$Installer..." -ForegroundColor Green
+        Start-Process -FilePath msiexec -Args "/i $Path\$Installer /quiet /qn /norestart" -Verb RunAs -Wait
+        Remove-Item $Path\$Installer
+    }
 
+}
 
 # Create a PowerShell ISE Shortcut on the Desktop
 $WshShell = New-Object -ComObject WScript.Shell
@@ -50,3 +62,4 @@ New-Item -ItemType Directory -Force -Path $allUsersDesktopPath
 $Shortcut = $WshShell.CreateShortcut("$allUsersDesktopPath\PowerShell ISE.lnk")
 $Shortcut.TargetPath = "$env:windir\system32\WindowsPowerShell\v1.0\PowerShell_ISE.exe"
 $Shortcut.Save()  
+}
